+++ CHANGELOG.md
@@ -19,14 +19,12 @@ The **Unreleased** section collects notes for unreleased changes and features, u
 
 ## Unreleased
 
+- [Added] changes to allow report record edit and create to work with arbitrary models
 
 ## [7.4.77] - 2021-11-19
 
-<<<<<<< HEAD
 - [Fixed] multiple bugfixes related to report criteria configuration and select_from_model
 - [Fixed] schema updated again for missing migration
-=======
->>>>>>> 7151fc46ea58da2565df974221d8ea5e8bd53c1c
 
 ## [7.4.76] - 2021-11-18